+++ src/pages/Home.tsx
@@ -1,6 +1,6 @@
 import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar, IonButtons, IonButton, IonSearchbar } from '@ionic/react';
 import ExploreContainer from '../components/ExploreContainer';
-import TextCard from '../components/TextCard';
+import TextCard from '../components/IntroductionTextCard';
 import './Home.css';
 
 const Home: React.FC = () => {
@@ -29,8 +29,13 @@ const Home: React.FC = () => {
               <IonSearchbar showClearButton="always" animated={true} placeholder="Enter country name here"></IonSearchbar>
             </div>
           </div>
+          <div className='row'>
+            <TextCard />
+          </div>
+          <div className='row'>
+            <ExploreContainer />
+          </div>
         </div>
-        <ExploreContainer />
       </IonContent>
     </IonPage>
   );
